backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
description = "Backend service for Project Management MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.122.0",
  "httpx>=0.28.1",
  "itsdangerous>=2.1.2",
  "python-dotenv>=1.2.1",
  "python-multipart>=0.0.20",
  "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
